requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
dnspython==2.5.0
elastic-transport==8.12.0
elasticsearch==8.12.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
frozenlist==1.4.1
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
libcst==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
orjson==3.9.14
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
strawberry-graphql==0.219.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.2.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
